refactor : redisService 패키지 변경 #170
