+++ src/main/java/com/devtraces/arterest/service/auth/AuthService.java
@@ -29,7 +29,7 @@ import java.util.Date;
 import java.util.Optional;
 import java.util.Random;
 
-import com.devtraces.arterest.service.user.RedisService;
+import com.devtraces.arterest.service.redis.RedisService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
