+++ src/test/java/com/devtraces/arterest/service/auth/AuthServiceTest.java
@@ -35,7 +35,7 @@ import java.util.List;
 import java.util.Optional;
 
 import com.devtraces.arterest.service.s3.S3Service;
-import com.devtraces.arterest.service.user.RedisService;
+import com.devtraces.arterest.service.redis.RedisService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
