+++ src/test/java/com/devtraces/arterest/service/user/UserServiceTest.java
@@ -4,6 +4,7 @@ import com.devtraces.arterest.common.type.UserSignUpType;
 import com.devtraces.arterest.controller.user.dto.response.*;
 import com.devtraces.arterest.model.follow.Follow;
 import com.devtraces.arterest.model.follow.FollowRepository;
+import com.devtraces.arterest.service.redis.RedisService;
 import com.devtraces.arterest.service.s3.S3Service;
 import com.devtraces.arterest.common.exception.BaseException;
 import com.devtraces.arterest.model.feed.FeedRepository;
